+++ src/fsCore/Common/Models/GroupCatchModel.cs
@@ -18,7 +18,7 @@ namespace Common.Models
         public WorldFish? WorldFish { get; set; }
         [JsonConstructor]
         public GroupCatch() { }
-        public GroupCatch(Guid userId, Guid groupId, string species, double weight, DateTime caughtAt, double length, double latitude, double longitude, string? description, Guid? id, DateTime? createdAt, byte[]? catchPhoto, Group? group, User? user, string? worldFishTaxocode, WorldFish? worldFish)
+        public GroupCatch(Guid userId, Guid groupId, string species, decimal weight, DateTime caughtAt, decimal length, decimal latitude, decimal longitude, string? description, Guid? id, DateTime? createdAt, byte[]? catchPhoto, Group? group, User? user, string? worldFishTaxocode, WorldFish? worldFish)
         {
             Id = id;
             Latitude = latitude;
