+++ src/fsCore/fsCore/ApiModels/SaveGroupCatchFormInput.cs
@@ -5,14 +5,14 @@ namespace fsCore.ApiModels
         public Guid? Id { get; init; }
         public Guid GroupId { get; init; }
         public string Species { get; init; }
-        public double Weight { get; init; }
-        public double Length { get; init; }
+        public decimal Weight { get; init; }
+        public decimal Length { get; init; }
         public string? Description { get; init; }
         public string CaughtAt { get; init; }
         public IFormFile? CatchPhoto { get; init; }
         public string? CreatedAt { get; init; }
-        public double Latitude { get; init; }
-        public double Longitude { get; init; }
+        public decimal Latitude { get; init; }
+        public decimal Longitude { get; init; }
         public string? WorldFishTaxocode { get; init; }
     }
 }
